[Calculations]: fixes in calc average item level

+++ src/server/game/Entities/Item/ItemTemplate.h
@@ -828,9 +828,10 @@ struct ItemTemplate
         return (Stackable == 2147483647 || Stackable <= 0) ? uint32(0x7FFFFFFF-1) : uint32(Stackable);
     }
 
-    float GetItemLevelIncludingQuality(uint32 level) const
+    uint32 GetItemLevelIncludingQuality(uint32 level) const
     {
-        float itemLevel = (float)level;
+        uint32 itemLevel = level;
+
         switch (Quality)
         {
             case ITEM_QUALITY_POOR:
@@ -848,6 +849,7 @@ struct ItemTemplate
             default:
                 break;
         }
+
         return itemLevel;
     }
